Rework the server side networking architecture
A `GameServer` class was introduced which will take care of handing messages and sending responses back.
We now have an accept thread for accepting incoming clients, and a new thread is created for each connected client to listen for incoming messages.
Received messages are added to a shared queue, to be processed by the game server.