GameServer: introduce `GAMELIST` and `GAMELIST_REQUEST` message types
These messages are used to simplify joining a game, without needing to know the exact game code to join.

The `GAMELIST_REQUEST` message is sent by the client to request a list of available games to join.
The `GAMELIST` message is sent as a reply to the above request, containing the game list.