Server: introduce `GameList` active games container
The `GameList` class will take care of managing active games.

Motivation: as we had introduced a `GAMELIST` message, we needed a way to track joinable games. By moving the game list management logic to a dedicated entity, we can achieve that in a transparent and efficient way. We are now able to update the join-able game list every time an operation is performed on any game.