Controller: fix double deletion of a game when leaving the server
The game code variable in controller was never reset after leaving a game. The client sending quit game messages would then use the old game code resulting in the server trying to remove a non existant game, and not disconnecting the client properly. We now save the old game code in a dedicated variable (used by the GUI to ease reconnections) and reset all game related variables when quitting.