Model: Big Character Card fix and rework
The game was using the "used" boolean value of the CharacterCardEnum, Enum are static so this value was shared throughout the entire application, hindering the multi game function. Now the "used" boolean is contained in the CharacterCard Object ,that had to become an Abstract class ( also incorporating some methods of its children).