Model: Another Big Character Card fix and rework
The "used" attribute wasn't working in the CharacterCard Object: when the server received an Apply Effect Action it would set "used" as true, this was a problem because the same reference was sent back to the clients with the changed value. This resulted in client removing a different amount of coins from the players than what had happened on the server with a consequent Desyncronization. Now to track if a card is used we use a usedCards EnumMap in PlayingField.